+++ src/test/java/org/cicirello/ds/BinaryHeapTests.java
@@ -495,7 +495,7 @@ public class BinaryHeapTests {
 			for (int j = 0; j < n; j++) {
 				pq.offer(elements[j], priorities[j]);
 			}
-			pq.change(elements[i], 1);
+			assertTrue(pq.change(elements[i], 1));
 			assertEquals(1, pq.peekPriority(elements[i]));
 			assertEquals(elements[i], pq.pollElement());
 			for (int j = 0; j < n; j++) {
@@ -511,7 +511,7 @@ public class BinaryHeapTests {
 			for (int j = 0; j < n; j++) {
 				pq.offer(elements[j], priorities[j]);
 			}
-			pq.change(elements[i], 100);
+			assertTrue(pq.change(elements[i], 100));
 			assertEquals(100, pq.peekPriority(elements[i]));
 			for (int j = 0; j < n; j++) {
 				if (i!=j) {
@@ -529,7 +529,7 @@ public class BinaryHeapTests {
 				for (int j = 0; j < n; j++) {
 					pq.offer(elements[j], priorities[j]);
 				}
-				pq.change(elements[i], p);
+				assertTrue(pq.change(elements[i], p));
 				assertEquals(p, pq.peekPriority(elements[i]));
 				int j = 0;
 				for (; j < n; j++) {
@@ -556,7 +556,7 @@ public class BinaryHeapTests {
 			for (int j = 0; j < n; j++) {
 				pq.offer(elements[j], priorities[j]);
 			}
-			pq.change(elements[i], priorities[i]);
+			assertFalse(pq.change(elements[i], priorities[i]));
 			assertEquals(priorities[i], pq.peekPriority(elements[i]));
 			for (int j = 0; j < n; j++) {
 				assertEquals(elements[j], pq.pollElement());
@@ -570,7 +570,7 @@ public class BinaryHeapTests {
 			for (int j = 0; j < n; j++) {
 				pq.offer(elements[j], priorities[j]);
 			}
-			pq.change("YYY", p);
+			assertTrue(pq.change("YYY", p));
 			assertEquals(p, pq.peekPriority("YYY"));
 			int j = 0;
 			for (; j < n; j++) {
@@ -1109,7 +1109,7 @@ public class BinaryHeapTests {
 			for (int j = 0; j < n; j++) {
 				pq.offer(elements[j], priorities[j]);
 			}
-			pq.change(elements[i], -1);
+			assertTrue(pq.change(elements[i], -1));
 			assertEquals(-1, pq.peekPriority(elements[i]));
 			assertEquals(elements[i], pq.pollElement());
 			for (int j = 0; j < n; j++) {
@@ -1125,7 +1125,7 @@ public class BinaryHeapTests {
 			for (int j = 0; j < n; j++) {
 				pq.offer(elements[j], priorities[j]);
 			}
-			pq.change(elements[i], -100);
+			assertTrue(pq.change(elements[i], -100));
 			assertEquals(-100, pq.peekPriority(elements[i]));
 			for (int j = 0; j < n; j++) {
 				if (i!=j) {
@@ -1143,7 +1143,7 @@ public class BinaryHeapTests {
 				for (int j = 0; j < n; j++) {
 					pq.offer(elements[j], priorities[j]);
 				}
-				pq.change(elements[i], -p);
+				assertTrue(pq.change(elements[i], -p));
 				assertEquals(-p, pq.peekPriority(elements[i]));
 				int j = 0;
 				for (; j < n; j++) {
@@ -1170,7 +1170,7 @@ public class BinaryHeapTests {
 			for (int j = 0; j < n; j++) {
 				pq.offer(elements[j], priorities[j]);
 			}
-			pq.change(elements[i], priorities[i]);
+			assertFalse(pq.change(elements[i], priorities[i]));
 			assertEquals(priorities[i], pq.peekPriority(elements[i]));
 			for (int j = 0; j < n; j++) {
 				assertEquals(elements[j], pq.pollElement());
@@ -1184,7 +1184,7 @@ public class BinaryHeapTests {
 			for (int j = 0; j < n; j++) {
 				pq.offer(elements[j], priorities[j]);
 			}
-			pq.change("YYY", -p);
+			assertTrue(pq.change("YYY", -p));
 			assertEquals(-p, pq.peekPriority("YYY"));
 			int j = 0;
 			for (; j < n; j++) {