+++ src/test/java/org/cicirello/ds/FibonacciHeapDoubleTests.java
@@ -522,12 +522,12 @@ public class FibonacciHeapDoubleTests {
 					assertEquals("T", e3);
 					assertEquals(n-3, pq.size());
 					if (!e1.equals(elements[i]) && !e2.equals(elements[i]) && !e3.equals(elements[i])) {
-						pq.change(elements[i], -1);
+						assertTrue(pq.change(elements[i], -1));
 						assertEquals(-1.0, pq.peekPriority(elements[i]), 0.0);
 						assertEquals(n-3, pq.size());
 					}
 					if (!e1.equals(elements[k]) && !e2.equals(elements[k]) && !e3.equals(elements[k])) {
-						pq.change(elements[k], -2);
+						assertTrue(pq.change(elements[k], -2));
 						assertEquals(-2.0, pq.peekPriority(elements[k]), 0.0);
 						assertEquals(n-3, pq.size());
 					}
@@ -597,12 +597,12 @@ public class FibonacciHeapDoubleTests {
 					assertEquals("T", e3);
 					assertEquals(n-3, pq.size());
 					if (!e1.equals(elements[i]) && !e2.equals(elements[i]) && !e3.equals(elements[i])) {
-						pq.change(elements[i], 95);
+						assertTrue(pq.change(elements[i], 95));
 						assertEquals(95.0, pq.peekPriority(elements[i]), 0.0);
 						assertEquals(n-3, pq.size());
 					}
 					if (!e1.equals(elements[k]) && !e2.equals(elements[k]) && !e3.equals(elements[k])) {
-						pq.change(elements[k], 102);
+						assertTrue(pq.change(elements[k], 102));
 						assertEquals(102.0, pq.peekPriority(elements[k]), 0.0);
 						assertEquals(n-3, pq.size());
 					}
@@ -644,7 +644,7 @@ public class FibonacciHeapDoubleTests {
 			for (int j = 0; j < n; j++) {
 				pq.offer(elements[j], priorities[j]);
 			}
-			pq.change(elements[i], 1);
+			assertTrue(pq.change(elements[i], 1));
 			assertEquals(1.0, pq.peekPriority(elements[i]), 0.0);
 			assertEquals(elements[i], pq.pollElement());
 			for (int j = 0; j < n; j++) {
@@ -660,7 +660,7 @@ public class FibonacciHeapDoubleTests {
 			for (int j = 0; j < n; j++) {
 				pq.offer(elements[j], priorities[j]);
 			}
-			pq.change(elements[i], 100);
+			assertTrue(pq.change(elements[i], 100));
 			assertEquals(100.0, pq.peekPriority(elements[i]), 0.0);
 			for (int j = 0; j < n; j++) {
 				if (i!=j) {
@@ -678,7 +678,7 @@ public class FibonacciHeapDoubleTests {
 				for (int j = 0; j < n; j++) {
 					pq.offer(elements[j], priorities[j]);
 				}
-				pq.change(elements[i], p);
+				assertTrue(pq.change(elements[i], p));
 				assertEquals(p, pq.peekPriority(elements[i]), 0.0);
 				int j = 0;
 				for (; j < n; j++) {
@@ -705,7 +705,7 @@ public class FibonacciHeapDoubleTests {
 			for (int j = 0; j < n; j++) {
 				pq.offer(elements[j], priorities[j]);
 			}
-			pq.change(elements[i], priorities[i]);
+			assertFalse(pq.change(elements[i], priorities[i]));
 			assertEquals(priorities[i], pq.peekPriority(elements[i]), 0.0);
 			for (int j = 0; j < n; j++) {
 				assertEquals(elements[j], pq.pollElement());
@@ -719,7 +719,7 @@ public class FibonacciHeapDoubleTests {
 			for (int j = 0; j < n; j++) {
 				pq.offer(elements[j], priorities[j]);
 			}
-			pq.change("YYY", p);
+			assertTrue(pq.change("YYY", p));
 			assertEquals(p, pq.peekPriority("YYY"), 0.0);
 			int j = 0;
 			for (; j < n; j++) {
@@ -745,7 +745,7 @@ public class FibonacciHeapDoubleTests {
 			}
 			String minElement = pq.pollElement();
 			if (!minElement.equals(elements[i])) {
-				pq.change(elements[i], priorities[i]-0.5);
+				assertTrue(pq.change(elements[i], priorities[i]-0.5));
 				assertEquals(priorities[i]-0.5, pq.peekPriority(elements[i]), 0.0);
 			}
 			for (int j = 0; j < n; j++) {
@@ -1176,12 +1176,12 @@ public class FibonacciHeapDoubleTests {
 					assertEquals("T", e3);
 					assertEquals(n-3, pq.size());
 					if (!e1.equals(elements[i]) && !e2.equals(elements[i]) && !e3.equals(elements[i])) {
-						pq.change(elements[i], 2);
+						assertTrue(pq.change(elements[i], 2));
 						assertEquals(2.0, pq.peekPriority(elements[i]), 0.0);
 						assertEquals(n-3, pq.size());
 					}
 					if (!e1.equals(elements[k]) && !e2.equals(elements[k]) && !e3.equals(elements[k])) {
-						pq.change(elements[k], 1);
+						assertTrue(pq.change(elements[k], 1));
 						assertEquals(1.0, pq.peekPriority(elements[k]), 0.0);
 						assertEquals(n-3, pq.size());
 					}
@@ -1251,12 +1251,12 @@ public class FibonacciHeapDoubleTests {
 					assertEquals("T", e3);
 					assertEquals(n-3, pq.size());
 					if (!e1.equals(elements[i]) && !e2.equals(elements[i]) && !e3.equals(elements[i])) {
-						pq.change(elements[i], -102);
+						assertTrue(pq.change(elements[i], -102));
 						assertEquals(-102.0, pq.peekPriority(elements[i]), 0.0);
 						assertEquals(n-3, pq.size());
 					}
 					if (!e1.equals(elements[k]) && !e2.equals(elements[k]) && !e3.equals(elements[k])) {
-						pq.change(elements[k], -95);
+						assertTrue(pq.change(elements[k], -95));
 						assertEquals(-95.0, pq.peekPriority(elements[k]), 0.0);
 						assertEquals(n-3, pq.size());
 					}
@@ -1298,7 +1298,7 @@ public class FibonacciHeapDoubleTests {
 			for (int j = 0; j < n; j++) {
 				pq.offer(elements[j], priorities[j]);
 			}
-			pq.change(elements[i], -1);
+			assertTrue(pq.change(elements[i], -1));
 			assertEquals(-1.0, pq.peekPriority(elements[i]), 0.0);
 			assertEquals(elements[i], pq.pollElement());
 			for (int j = 0; j < n; j++) {
@@ -1314,7 +1314,7 @@ public class FibonacciHeapDoubleTests {
 			for (int j = 0; j < n; j++) {
 				pq.offer(elements[j], priorities[j]);
 			}
-			pq.change(elements[i], -100);
+			assertTrue(pq.change(elements[i], -100));
 			assertEquals(-100.0, pq.peekPriority(elements[i]), 0.0);
 			for (int j = 0; j < n; j++) {
 				if (i!=j) {
@@ -1332,7 +1332,7 @@ public class FibonacciHeapDoubleTests {
 				for (int j = 0; j < n; j++) {
 					pq.offer(elements[j], priorities[j]);
 				}
-				pq.change(elements[i], -p);
+				assertTrue(pq.change(elements[i], -p));
 				assertEquals(-p, pq.peekPriority(elements[i]), 0.0);
 				int j = 0;
 				for (; j < n; j++) {
@@ -1359,7 +1359,7 @@ public class FibonacciHeapDoubleTests {
 			for (int j = 0; j < n; j++) {
 				pq.offer(elements[j], priorities[j]);
 			}
-			pq.change(elements[i], priorities[i]);
+			assertFalse(pq.change(elements[i], priorities[i]));
 			assertEquals(priorities[i], pq.peekPriority(elements[i]), 0.0);
 			for (int j = 0; j < n; j++) {
 				assertEquals(elements[j], pq.pollElement());
@@ -1373,7 +1373,7 @@ public class FibonacciHeapDoubleTests {
 			for (int j = 0; j < n; j++) {
 				pq.offer(elements[j], priorities[j]);
 			}
-			pq.change("YYY", -p);
+			assertTrue(pq.change("YYY", -p));
 			assertEquals(-p, pq.peekPriority("YYY"), 0.0);
 			int j = 0;
 			for (; j < n; j++) {
@@ -1399,7 +1399,7 @@ public class FibonacciHeapDoubleTests {
 			}
 			String maxElement = pq.pollElement();
 			if (!maxElement.equals(elements[i])) {
-				pq.change(elements[i], priorities[i]+0.5);
+				assertTrue(pq.change(elements[i], priorities[i]+0.5));
 				assertEquals(priorities[i]+0.5, pq.peekPriority(elements[i]), 0.0);
 			}
 			for (int j = 0; j < n; j++) {